% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventSequenceGraph.R
\name{.buildEventNameLookup}
\alias{.buildEventNameLookup}
\title{Build a named lookup vector: character(comboId) -> "EventA + EventB"}
\usage{
.buildEventNameLookup(codesLong, generationSet)
}
\description{
Build a named lookup vector: character(comboId) -> "EventA + EventB"
}
\keyword{internal}
