% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OdysseusPathwayModule-package.R
\docType{package}
\name{OdysseusPathwayModule-package}
\alias{OdysseusPathwayModule}
\alias{OdysseusPathwayModule-package}
\title{OdysseusPathwayModule: Cohort Pathway Analysis with Pre-Index Event Support}
\description{
Provides cohort pathway analysis for Observational Medical Outcomes Partnership (OMOP) Common Data Model databases, including both standard (post-index) and pre-index pathway analyses. The pre-index analysis identifies sequences of events occurring in a lookback window before the target cohort index date. Built on the 'CohortPathways' analysis framework originally developed by Christopher Knoll and the Observational Health Data Sciences and Informatics community through 'WebAPI'. Methodological background and the originating implementation are described in \url{https://github.com/OHDSI/CohortPathways}.
}
\author{
\strong{Maintainer}: Alexander Alexeyuk \email{AlexanderAlexeyuk@gmail.com}

}
\keyword{internal}
